Maintain the intended design language
The one design fits all of many current plamo apps fits
many of them but not all of them, and top toolbars on mobile,
while are not always avoidable, they do have problems which
have not been adressed.
This is a small and "uninportnant" application, in which is
worth experimenting something else/better.

In this i really want to keep the initial core design principles
of Kirigami which i think they are still valid:
* Bottom-heavy controls
* No top toolbars on mobile as in this kind of app can be easily avoided
* Floating action button for the main action
* gesture based page switching
* when there is enough space, show all the pages at once